test: use ChromeHeadlessCI, for compatibility with Docker
In GH Actions, --no-sandbox is not longer needed, so ChromeHeadless
works fine. Ditto on Windows. But in my Docker container on Windows, I
had to use ChromeHeadlessCI for `npm run test:once` to work. So let's
stick to that for maximum compatibility everywhere, since the CI version
with --no-sandbox works reliably everywhere.