fix: just say MP3 and WAV as supported audio formats, but also allow M4A and WEBM
.webm is an obscure format, most user will not know what it is. We have
to support it since the Save button generates it on Windows, but it's
enough to support it. Having it in the `accept` list means it'll be
avaiable to select so it should not be (too) confusing.

I'm also adding .m4a to the `accept` list (without mentioning it in the
text) because on Windows, that's the default format you get when you
record, and it's the format of my sample in the Studio.git repo that
found its way everywhere, so we really should allow selecting it!

Fixes #161