fix(ci): the bundle is not actually stable, cannot validate with cmp
git status --porcelain demands exact equality of the committed bundle
and the one built in CI, but the minification and optimization process
is not completely deterministic, so it's not a reliable test.

We don't actually use the bundle in the published packages, and we in
fact rebuild the bundle automatically as part of the publish.yml
workflow that deploys to gh-pages and production, so this is not useful
test at all!