fix(publish): fix hyperlinks on npmjs READMEs
The hyperlink to `../web-component` in ngx readme didn't work on npmjs.
It linked to https://www.npmjs.com/package/web-component instead
of https://www.npmjs.com/package/@readalongs/web-component.
It's a nuisance that relative URLs in README work differently on npmjs
and github, but we can work-around it ourselves.

And the hyperlink to test data in web-c readme was completely broken
since we don't publish the test data to npmjs at all.

Fix both in the release workflow using sed between build and publish.

Fixes #311