ci: exercise our tests and scripts on Windows too
This is more expensive, though, and not critical, so we don't want to
run these tests on every push in every PR, only on the actual push to
main.

Known limitation: "npx nx bundle web-component" only works in Git Bash on
Windows, not in cmd or PowerShell, so don't exercise it here, it'll just
fail every time.

Second limitation: we end up installing stuff twice, once for Cypress,
once for the rest of the processes. I'm not sure how to avoid that.
I leave it as future work to optimize this out.