build: make bundle compatible with Windows and test it in CI (#344)
In CI, let's this npx nx bundle web-component on both Windows and Linux, and on
Linux let's display the diff in a way that we can see the bits that changed
even though it's minified.

To make bundle compatible with Windows, I turned the one liner in `package.json`
into a script that is much easier to read and maintain.

(cherry picked from commit 1690f7e89e726c0c38680edbe9f6a7d8b876a78a)