build: do not commit browserslist-db updates, it is just noise
The strategy I implemented in #348 to always update the browserslist-db
on install *and commit the results* creates a lot of noise: a lot of
similar commits, and we still want to update this db in deploy anyway,
so it's not the right strategy.

This commit removes the code that commits the changes, and instead
updates the db in every workflow that tests or deploys Studio-Web.

Furthermore, Del is going to add a build-time dependency that will do
this update automatically when you build or serve locally, so we do the
same everywhere, automatically, and reasonably quietly.